Set styles for image size to account for new image
I realized the original images I was using only had Waldo and none of
the other characters. I found a newer version of the image with all the
other characters, but the image is massive. Hardcoding the default size
to match my 2k monitor. This will ensure the coordinates still align
even if the user zooms in or out.